--- src/app/models/albums.ts
@@ -1,6 +1,0 @@
-import { Album } from './album';
-
-export class Albums {
-  data: Album[];
-  next: string;
-}
