--- src/app/models/artists.ts
@@ -1,6 +1,0 @@
-import { Artist } from './artist';
-
-export class Artists {
-  data: Artist[];
-  next: string;
-}
