--- src/app/models/tracks.ts
@@ -1,6 +1,0 @@
-import { Track } from './track';
-
-export class Tracks {
-  data: Track[];
-  next: string;
-}
